pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dinosaur"]
package-data = {dinosaur = ["data/*"]}

[project]
name = "dinosaur"
version = "1.3.5"  # keep sync with __init__.py
description = "Differentiable Dynamics for Global Atmospheric Modeling 🦖"
authors = [
    {name = "Google LLC", email = "noreply@google.com"},
]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
# TODO(shoyer): thin these down
dependencies = [
    "jax",
    "jaxlib",
    "numpy",
    "pint",
    "scipy",
    "scikit-learn",
    "tree-math",
    # these packages below are only required for xarray-utils, and really
    # should not be required by default.
    "dask",
    "fsspec",
    "pandas",
    "xarray",
    # "xarray-beam",  # TODO(shoyer): figure out how to install Beam in CI
    "xarray-tensorstore",
]

[project.urls]
source = "https://github.com/neuralgcm/dinosaur"

[project.optional-dependencies]
tests = [
    "absl-py",
    "chex",
    "pytest",
]
